homepage handlebars done, signup handlbars done. Addition of signup.js to public folder
