+++ .env
@@ -1,4 +1,4 @@
-DB_NAME=stagename_db
-DB_PASSWORD=Abc123
-DB_USER=root
+DB_NAME='stagename_db'
+DB_PASSWORD='Arapiles052021.'
+DB_USER='root'
