+++ public/css/style.css
@@ -6,14 +6,26 @@
     --alternatedarkcolor: #6A544D;
 }
 
-header {
-    background-color: var(--mainthemecolor);
-    height: 20rem
+body {
+    overflow-x: hidden;
 }
 
-h1,
+header {
+    background-color: var(--mainthemecolor);
+    height: 10rem;
+    justify-content: center;
+    align-items: center;
+}
+
+h1 {
+    color: var(--textcolor);
+    font-weight: bold;
+    text-align: center;
+    margin-top: 2.5rem;
+}
+
 h2,
-h3 {
+h6 {
     color: var(--textcolor);
     font-weight: bold;
     text-align: center;
@@ -24,7 +36,6 @@ main {
     width: 100%;
     background-color: var(--mainthemecolor);
     color: var(--textcolour);
-    overflow-x: hidden;
     margin: 0;
 }
 
@@ -50,6 +61,8 @@ footer {
     justify-content: center;
     align-items: center;
     text-align: center;
+    margin-top: 1.5rem;
+    width: 100%
 }
 
 .btn {
