+++ views/layouts/main.handlebars
@@ -1,9 +1,7 @@
 <html lang="en">
-
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Music Fwends</title>
     <link
       rel="stylesheet"
       href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/css/bootstrap.min.css"
@@ -11,6 +9,8 @@
       crossorigin="anonymous"
     />
     <link rel="stylesheet" href="/css/style.css" />
+
+    <title>Music Fwends</title>
   </head>
 
   <body>
